business logic for tickets implemented

+++ ticket_catalogue_service/src/main/kotlin/it/polito/wa2/ticket_catalogue_service/entities/Ticket.kt
@@ -19,5 +19,5 @@ data class Ticket(
         val maxAge: Int?,
         val start_period: LocalDateTime?,
         val end_period: LocalDateTime?,
-        val duration: Int? = null
+        val duration: Long? = null
 )
